Correction of the "solveDay21Puzzle2" function's return data type in the solution for day 21 - puzzle 1

+++ AdventOfCode2020/day21/day21puzzle1.cpp
@@ -69,7 +69,7 @@ void readInputsDay21Puzzle1(vector<food_T>& foodList, set<string>& allergensSet)
 	inFile.close();
 }
 
-long long solveDay21Puzzle1(vector<food_T> foodList, set<string> allergensSet)
+int solveDay21Puzzle1(vector<food_T> foodList, set<string> allergensSet)
 {
 	// Create the "allergen2Ingredients" set which contains for each allergen all the possible ingredients which can contain it
 	map<string, set<string>> allergen2Ingredients;
